making changes in CSS
following the following presenations:
11/28/2014
https://speakerdeck.com/vjeux/react-css-in-js
12/14/2014
https://speakerdeck.com/vjeux/react-css-in-js-react-france-meetup
