+++ webpack.config.js
@@ -9,6 +9,7 @@ module.exports = {
   },
   module: {
     loaders: [
+      { test: /\.js$/, loader: 'jsx-loader?harmony' },
       { test: /\.jsx$/, loader: 'jsx-loader?harmony' },
       { test: /\.less$/, loader: 'style-loader!css-loader!less-loader' }, // use ! to chain loaders
       { test: /\.css$/, loader: 'style-loader!css-loader' },
